refactor/filter: Move CombinedData into utils.search_engine.source_data
