+++ utils/search_engine/source_data.py
@@ -1,4 +1,11 @@
+from dataclasses import dataclass
 from typing import Final
+
+import polars as pl
+
+from utils.search_engine.entity import Playlist, PlaylistTrack, Track, TrackAdjacent, TrackLyrics
+from utils.search_engine.entity_base import PolarsLazyFrame
+from utils.search_engine.filters import PlaylistSet, PlaylistTrackSet, TrackLyricsSet, TrackSet
 
 
 PLAYLIST_DATA_FILE: Final = 'data_playlist_metadata.parquet'
@@ -12,3 +19,43 @@ COUNTRY_DATA_FILE: Final = 'data_countries.parquet'
 
 TRACK_DUPLICATES_DATA_FILE: Final = 'data_song_duplicates.parquet'
 TRACK_CANONICAL_DATA_FILE: Final = 'data_song_canonical.parquet'
+
+
+@dataclass(kw_only=True)
+class CombinedData:
+    """Holder for the different underlying data sources."""
+
+    playlists: PolarsLazyFrame[Playlist]
+    playlist_tracks: PolarsLazyFrame[PlaylistTrack]
+    tracks: PolarsLazyFrame[Track]
+    tracks_adjacent: PolarsLazyFrame[TrackAdjacent]
+    track_lyrics: PolarsLazyFrame[TrackLyrics]
+    countries: list[str]
+
+    @property
+    def all_playlists(self) -> PlaylistSet:
+        return PlaylistSet(self.playlists, None, self.playlists, is_filtered=False)
+
+    @property
+    def all_playlist_tracks(self) -> PlaylistTrackSet:
+        return PlaylistTrackSet(self.playlist_tracks, is_filtered=False)
+
+    @property
+    def all_tracks(self) -> TrackSet:
+        return TrackSet(self.tracks, is_filtered=False)
+
+    @property
+    def all_track_lyrics(self):
+        return TrackLyricsSet(self.track_lyrics, is_filtered=False)
+
+    @staticmethod
+    def load_from_files():
+        """Load the pre-generated data from the Parquet files."""
+        return CombinedData(
+            playlists=pl.scan_parquet(PLAYLIST_DATA_FILE),
+            playlist_tracks=pl.scan_parquet(PLAYLIST_TRACKS_DATA_FILE),
+            tracks=pl.scan_parquet(TRACK_DATA_FILE),
+            tracks_adjacent=pl.scan_parquet(TRACK_ADJACENT_DATA_FILE),
+            track_lyrics=pl.scan_parquet(TRACK_LYRICS_DATA_FILE),
+            countries=pl.read_parquet(COUNTRY_DATA_FILE)['country'].to_list(),
+        )
